requirements.txt
imagehash
Pillow
pynput
schedule
